bot/requirements.txt
pybit>=5.6.2
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
loguru==0.7.3
python-dotenv>=0.19.0
python-dateutil>=2.8.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-telegram-bot==20.7
schedule==1.2.0
configparser==6.0.0
requests
scikit-learn>=1.3.0